requirements.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.12
djangorestframework-simplejwt==5.2.2
django_filter==23.1
channels==3.0.4
Pillow==9.5.0

PyYAML==6.0
uritemplate==4.1.1

# for visual studio code
autopep8
pylint
pylint-django
pylint-plugin-utils
flake8
pyflakes
